Introduce ability to define config values with Builder
When you set up builder_class for the option then you can have embedded
DSL. So something like

Doorkeeper.configure do
	attribute_method do
		#inner DSL for given attribute
	end
end

is possible.

The builder for attributes should act just like config attribute - it
takes the block for the initializer and then the method build returns
the value for the attribute.