Introduce pre authorization - WIP
PreAuthorization class is responsible for validating all parameters.
It can be used to display the authorization message, create the
grant/token or deny the request.

It returns a response which is redirectable or renderable. Since
OAuth spec requires that some error responses have to be returned
to the client in the redirect URI, this logic gets complicated

Still a bunch of duplicated code has to be removed.

Ideally:

- CodeRequest, TokenRequest
  - success: returns CodeResponse. Is always redirectable and can
    redirect either to test url or client's redirect uri
  - error: returns ErrorResponse. Needs to be splitted into
    CodeErrorResponse and TokenErrorResponse, since token errors
    are never redirectable whilst code don't