Add a facade to more easily get token from request
If you wish to use Doorkeeper outside of a Rails controller you must dig
deep into the internals of Doorkeeper to extract a token from an
incoming request. I needed to do this, for example, in a custom route
constraint. This PR adds a small facade which makes doing this simpler.