Limit scopes in app scopes for client credentials
With following configuration:

* `doorkeeper.configuration.default_scopes` set to `public`.
* `doorkeeper.configuration.optional_scopes` set to `x, y, z`.
* A registered application with `scopes` set to `x, y`.

When app requests an access token using client credential flow, the
results are:

* if `params[:scope]` is not present, result token will have scopes `public`
* if `params[:scope]` is set to `public`, server will deny the request
  because `public` is not in `app1.scopes`

This is strange, since when requesting `public` scope, it says the scope
is invalid. But when not requesting any specific scope, it grants a
`public` scope.

When `params[:scope]` is not present, server should either deny the
request or grant all scopes belonging to the client (in this case should
be `x, y`), or grant it the default scope only if default scope belongs
to the client.

[fixes #593]