Change the doorkeeper forcing timezones to be UTC.
This was initially a change for the sake of the HoundCI, but it is also just A Good Idea™.

Forcing UTC should not cause any issues as the way two times are compared takes the timezone into account anyway.

Here are a few articles regarding the best practice of Time in Ruby.
https://robots.thoughtbot.com/its-about-time-zones
https://robots.thoughtbot.com/a-case-study-in-multiple-time-zones
http://rails-bestpractices.com/posts/2014/10/22/use-time-zone-now-instead-of-time-now/
http://danilenko.org/2012/7/6/rails_timezones/