Rename user factory to doorkeeper_testing_user
Developers of Rails applications love to include the factories from the doorkeeper gem into their applications because the factories make it easy to do feature testing of the API, stubbing out the relevant doorkeeper models.

However, there is a name conflict on the user factory (which is a global namespace), so that developers cannot include `<path_to_doorkeeper>/spec/factories.rb` if their own factories.rb includes a :user factory (which most do).

This change simply renames the internal factory (which is only used in 1 test anyways) to a scoped name that is much less likely to conflict.  It does not change which class the factory is modeled off of (still the user.rb in the dummy spec app).