Support redirect_uri's with registered query param(s)
This amends the URIChecker to allow authorization for requests that contain query params. The query params all must be defined in the client app's #redirect_uri and all defined params must be present. Their order does not matter.

If there is no query param registered in the app's #redirect_uri then the previous behaviour occurs: stripping the query param from the request and comparing the URLs without it.

Fixes #1050