Uses `expires_in` on token show endpoint for consistency
Creating an access token yields a response body with the parameter "expires_in" which is recommended as part of the OAuth2 spec. It makes sense to be consistent with this parameter on the info/show token equivalent endpoint instead of calling it "expires_in_seconds"

This makes consumers of Doorkeeper write less code since the response from both POST /oauth/token and GET /oauth/token/info are more uniform.

Rename scopes to scope for consistency with OAuth spec terminology

Add NEWS entry