Implement optional hashing of tokens and secrets
There have been previous requests with regards to encryption or hashing
OAuth secrets in the database to avoid storing them in plain texts.
Primarily, these discussions have taken place in #320 and #675 (at
least, that's where I took note of them).

For us, adaptation of doorkeeper is limited due to this exact storing of
plain text tokens, especially longlived refresh tokens as well as client
secrets in the database. Different from some of the commenters in the
above discussion, we do believe hashing alone is sufficient to address
the immediate threat of leaking tokens and secrets, e.g., from an
unencrypted backup.

Encryption does not provide much of an advantage if an attacker is able
to get access to the resource server and leak the encryption key.

This PR thus implements an *optional* hashing filter configurable
through the initializer that:

1. Hashes access_tokens, refresh_tokens, access_grants and application
client_secrets. Hashing method is SHA256 for tokens, BCrypt password
hashing for application secrets. Both methods are configurable, but
deliberately not through the end-user initializer.

2. Allow models to maintain the volatile plaintext value for one-time
communication to the user (e.g., showing the application, token in the
same request).

3. Alters lookups to allow finding input tokens through either
their hashed value, or the original plaintext as fallback. This allows
users to switch on this hashing option without losing any old
authenticated clients. They may want to reduce TTL of their refresh
tokens to ensure all clients will have hashed secrets persisted in the
near future.

Arguments against some of the discussions:

1. Using bcrypt for everything. Since the original plaintext token is a cryptographically
secure random string, we do not gain much by salting the values for tokens. The
UniqueToken generator of doorkeeper has 128-bit entropy by default, and
can be extended. However, this implementation allows users to use bcrypt or
whatever they prefer, it is just my opinion that the extra overhead is
useless.

2. Not storing tokens at all whereever possible. This was a fascinating
read, but out of my scope of what I'm capable to implement and would
be completely orthogonal to what doorkeeper currently does. I for one
like having the tokens around for auditing purposes.