Try to load bcrypt for application hashing, but add fallback
Currently bcrypt was an implicit dependency of the token hashing added
to bcrypt without explicitly adding it to the gemspec.

Since bcrypt hashing on applications was merely a choice without much
implied security, we should not push this requirement on all users of
doorkeeper.

Instead, this PR will try to configure `Doorkeeper::Application` with
bcrypt secret hashing and fall back to the default SHA256 digest if it
was not found.

A documentation section was added to the README that introduces token
hashing and also documents how to add bcrypt.

Fixes #1190