Make revocation of old access tokens optional
Revocation of access tokens obtained via client credentials
was (re-)introduced in Doorkeeper 5.2. However, this is incompatible
with clients that are distributed over multiple physical machines,
that use the same set of credentials. In this case it could happen
that one machine invalidates the credentials fetched by another
machine.

Since there seems to be a use case/demand for the automated revocation,
I decided to turn it into a configuration option.