Add support for configuring a custom application secret generator
Right now the application secret generator can't be configured and the
only way to ensure that application secrets have an identifiable prefix
is by overwriting `renew_secret`.

Generating identifiable tokens and secrets is important and relevant for
things like the GitHub Secret Scanning program and we use this already
for other things like regular OAuth tokens where we overwrite them with
a prefix so they are clearly identifiable for secret scanning.

The changes here also allow such an override for the application secret.