Use validates_with to validate redirect_uri
ActiveModel's validates method uses string manipulation, in particular
camelize, to identify the class constant to be used for validation. If
inflections have been modified or added, it may attempt to load a class
that does not exist. By using validates_with and passing the exact class
to be used, we can avoid this issue.