DRY out AuthorizationsController#render_success
In our own deployment, we have some custom exception handling around
the `create`` method. By calling the more "public" `create` method in
`render_success`, we get to have a _single_ code path that also supports
`skip_authorization?` and `can_authorize_response?`. The code on L35
is identical to L16.  That suggests identical intent, but not always.
This might be coincidental, in which case, the pull request should be
declined.